refactor(share): extract shared crit-share.js module for code-review + preview
The Share flow lived entirely in app.js (code-review only). Preview mode runs
live-mode.js chrome, so the Share button — though shown — was wired to nothing.

Extract the whole flow into frontend/crit-share.js (IIFE + dual-export
window.crit.share), a factory that owns all share state and wires the #shareBtn
click handler. Both modes consume it via injected adapters/options:

- app.js builds options and calls window.crit.share.create(...).reveal()
- live-mode.js installShareController() (preview only) reads /api/config and
  creates the controller with reviewType:'preview', canShare:true.

Transport, not protocol: every endpoint is byte-for-byte the pre-extraction
app.js code; the proxy_auth relay still hits the same crit-web APIs. Preview's
only branch is the payload-builder endpoint (/api/share/preview-payload).

Also:
- index.html: add crit-share.js + crit-icons.js to liveDeps; mark <html> with
  crit-mode-preview so style-live.css un-hides #shareBtn in preview (design/live
  stay hidden — nothing to publish behind a running app).
- live-mode.js: use canonical window.crit.icons.ICON_CLIPBOARD/ICON_CHECK_SMALL
  for the share modal copy buttons instead of a hand-rolled filled-clipboard
  glyph, so preview matches code-review exactly.
- frontend/__tests__/crit-share.test.js: 8 node --test cases for the module.
